Rename SQLService and GoGService files for consistency
Updated file names and corresponding includes to ensure consistency with naming conventions. These changes improve code readability and align with the project's naming standards.
